Correct the checksum-neutral findings of the whole-codebase review (#588)
The whole-codebase review of 2026-09-05 found these defects. This pull
request holds the fixes that keep the checksums identical. The fixes
that change the results come in separate pull requests.

- `input.cc`: `setup_phixs_list()` gives the ground level the slot of
its own ion (PR 554 stopped a dataset with two equal ground thresholds,
e.g. Er I and Tl I in the incfloers data); a phixs row whose upper ion
is not the next ion stops the run; a malformed integer line of
`input.txt` stops the run; an ion of an NLTE element with no excited
NLTE level stops the run; `num_lte_timesteps` must not exceed
`FIRST_NLTE_RADFIELD_TIMESTEP` with the multibin radiation field; one
`f_lu` formula with its finiteness check
- `nonthermal.cc`: a fixed-width parse of `auger-km1993-table2.txt`. A
five-digit ionisation potential touched the shell number, so a model
with Cu XX or Zn VIII stopped
- `nltepop.cc`: `nltepop_reset_element()` fills a subspan, so a TESTMODE
build no longer forms a past-the-end subscript; the NLTE output skips an
element with no mass in the cell (it wrote -rho, -inf, and nan)
- `grid.cc`: the electron fraction is negative until the input sets it,
and the TANAKA2020 grey opacity stops on a cell without one (it used 0.4
and the 0.96 fallback); the four log calls of `boundary_distance()` use
`MY_IF_HOST()` (a GPU build with `FORCE_SPHERICAL_ESCAPE_SURFACE` did
not compile); `ptrdiff_t` indices in the ion range accessors; the rank
range getters lose `[[gnu::pure]]`, because the first call sets up the
ranges and writes a file
- `rpkt.cc`: the expansion-opacity retrace keeps the line position of
the packet at the start of the path, so the line that emitted the packet
is excluded. No CI test reaches the retrace
- `gammapkt.cc`: checked reads of the gamma spectrum table; the discrete
deposition tally also for a parameterised gamma scheme with gamma
products
- `update_packets.cc`: no cooling sentinel for an ion with no cooling
term (the offset of such a last ion is one past the slot)
- `radfield.cc`: a `ptrdiff_t` index in the restart write
- `macroatom.cc`: a mutex for `macroatom_file`
- `mpi_logging.h`: no log line in the destructor of `MPI_shared_array`
(it ran during static destruction)
- `ltepop.cc`: the normalisation test is `isfinite`, not `== 0`
- `globals.h`: three `LevelAutoion` members that nothing reads are
removed
- `Makefile`: `REPRODUCIBLE=ON` overrides a command-line `FASTMATH`; a
bad value of `FASTMATH` or `OPTIMIZE` stops; `sn3dwhole` depends on the
headers
- `tests/setup_*.sh`: `curl` stops on an HTTP error and retries, so an
error page is not kept as the archive
- log and comment text

The checksums stay identical.


---------

+++ radfield.cc
@@ -1088,8 +1088,8 @@ void write_restart_data(FILE* gridsave_file) {
     const int bfestimcount = static_cast<int>(globals::bfestim_nu_edge.size());
     fprintf(gridsave_file, "%d\n", bfestimcount);
 
-    for (int nonemptymgi = 0; nonemptymgi < grid::get_nonempty_npts_model(); nonemptymgi++) {
-      fprintf(gridsave_file, "%d\n", nonemptymgi);
+    for (ptrdiff_t nonemptymgi = 0; nonemptymgi < grid::get_nonempty_npts_model(); nonemptymgi++) {
+      fprintf(gridsave_file, "%td\n", nonemptymgi);
       for (int i = 0; i < bfestimcount; i++) {
         fprintf(gridsave_file, "%a ", prev_bfrate_normed[(nonemptymgi * bfestimcount) + i]);
       }
